Add files via upload
Here we deploy a computer vision model trained to detect common objects in pictures. Deploying a model is one of the last steps in a prototypical machine learning lifecycle. However, we thought it would be exciting to get you to deploy a model right away. This lab uses a pretrained model called YOLOV3. This model is very convenient for two reasons: it runs really fast, and for object detection it yields accurate results.

The sequence of steps/tasks to complete in this lab are as follow:

1.Inspect the image data set used for object detection.
2.Take a look at the model itself.
3. Deploy the model using fastAPI.